+++ relay-config/src/config.rs
@@ -10,6 +10,7 @@ use std::str::FromStr;
 use std::time::Duration;
 
 use failure::{Backtrace, Context, Fail};
+use serde::de::{Unexpected, Visitor};
 use serde::{de::DeserializeOwned, Deserialize, Deserializer, Serialize, Serializer};
 
 use relay_auth::{generate_key_pair, generate_relay_id, PublicKey, RelayId, SecretKey};
@@ -983,13 +984,89 @@ impl Default for OutcomeAggregatorConfig {
     }
 }
 
+/// Determines how to emit outcomes.
+/// For compatibility reasons, this can either be true, false or AsClientReports
+#[derive(Copy, Clone, Debug, PartialEq, Eq)]
+
+pub enum EmitOutcomes {
+    /// Do not emit any outcomes
+    None,
+    /// Emit outcomes as client reports
+    AsClientReports,
+    /// Emit outcomes as outcomes
+    AsOutcomes,
+}
+
+impl EmitOutcomes {
+    /// Returns true of outcomes are emitted via http, kafka, or client reports.
+    pub fn any(&self) -> bool {
+        !matches!(self, EmitOutcomes::None)
+    }
+}
+
+impl Serialize for EmitOutcomes {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        // For compatibility, serialize None and AsOutcomes as booleans.
+        match self {
+            Self::None => serializer.serialize_bool(false),
+            Self::AsClientReports => serializer.serialize_str("as_client_reports"),
+            Self::AsOutcomes => serializer.serialize_bool(true),
+        }
+    }
+}
+
+struct EmitOutcomesVisitor;
+
+impl<'de> Visitor<'de> for EmitOutcomesVisitor {
+    type Value = EmitOutcomes;
+
+    fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
+        formatter.write_str("true, false, or 'as_client_reports'")
+    }
+
+    fn visit_bool<E>(self, v: bool) -> Result<Self::Value, E>
+    where
+        E: serde::de::Error,
+    {
+        Ok(if v {
+            EmitOutcomes::AsOutcomes
+        } else {
+            EmitOutcomes::None
+        })
+    }
+
+    fn visit_str<E>(self, v: &str) -> Result<Self::Value, E>
+    where
+        E: serde::de::Error,
+    {
+        if v == "as_client_reports" {
+            Ok(EmitOutcomes::AsClientReports)
+        } else {
+            Err(E::invalid_value(Unexpected::Str(v), &"as_client_reports"))
+        }
+    }
+}
+
+impl<'de> Deserialize<'de> for EmitOutcomes {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        deserializer.deserialize_any(EmitOutcomesVisitor)
+    }
+}
+
 /// Outcome generation specific configuration values.
 #[derive(Serialize, Deserialize, Debug)]
 #[serde(default)]
 pub struct Outcomes {
     /// Controls whether outcomes will be emitted when processing is disabled.
     /// Processing relays always emit outcomes (for backwards compatibility).
-    pub emit_outcomes: bool,
+    /// Can take the following values: false, "as_client_reports", true
+    pub emit_outcomes: EmitOutcomes,
     /// Controls wheather client reported outcomes should be emitted.
     pub emit_client_outcomes: bool,
     /// The maximum number of outcomes that are batched before being sent
@@ -1008,7 +1085,7 @@ pub struct Outcomes {
 impl Default for Outcomes {
     fn default() -> Self {
         Outcomes {
-            emit_outcomes: false,
+            emit_outcomes: EmitOutcomes::AsClientReports,
             emit_client_outcomes: true,
             batch_size: 1000,
             batch_interval: 500,
@@ -1487,8 +1564,11 @@ impl Config {
     ///
     /// This is `true` either if `outcomes.emit_outcomes` is explicitly enabled, or if this Relay is
     /// in processing mode.
-    pub fn emit_outcomes(&self) -> bool {
-        self.values.outcomes.emit_outcomes || self.values.processing.enabled
+    pub fn emit_outcomes(&self) -> EmitOutcomes {
+        if self.processing_enabled() {
+            return EmitOutcomes::AsOutcomes;
+        }
+        self.values.outcomes.emit_outcomes
     }
 
     /// Returns whether this Relay should emit client outcomes
@@ -1855,5 +1935,26 @@ cache:
         let values: ConfigValues = serde_yaml::from_str(yaml).unwrap();
         assert_eq!(values.cache.envelope_buffer_size, 1_000_000);
         assert_eq!(values.cache.envelope_expiry, 1800);
+    }
+
+    #[test]
+    fn test_emit_outcomes() {
+        for (serialized, deserialized) in &[
+            ("true", EmitOutcomes::AsOutcomes),
+            ("false", EmitOutcomes::None),
+            ("\"as_client_reports\"", EmitOutcomes::AsClientReports),
+        ] {
+            let value: EmitOutcomes = serde_json::from_str(serialized).unwrap();
+            assert_eq!(value, *deserialized);
+            assert_eq!(serde_json::to_string(&value).unwrap(), *serialized);
+        }
+    }
+
+    #[test]
+    fn test_emit_outcomes_invalid() {
+        assert!(matches!(
+            serde_json::from_str::<EmitOutcomes>("asdf"),
+            Err(_)
+        ));
     }
 }