+++ relay-metrics/benches/aggregator.rs
@@ -131,6 +131,7 @@ fn bench_insert_and_flush(c: &mut Criterion) {
         bucket_interval: 1000,
         initial_delay: 0,
         debounce_delay: 0,
+        ..Default::default()
     };
 
     let flush_receiver = TestReceiver.start().recipient();